integration: disable two x64 tests
Disables two tests that rely on initialization
of local variable in order to run recompiled
code successfuly. The solution for this situation
is at this point questionable as original
code does not have initialized variables as well.

If this test should have initialized variable
than it is required to detect that this variables
are 32 bit by stack analysis.