r2pluign: provide better functionality testing
Currenty R2 stores computer-dependant information
that is not suitable for computer-independent testing.

This commit brings changes to testing - provides
way to pass specific commands that will be executed
as initial commands after R2 loads the input binary.

As R2 projects may be improved in the future,
the functionality to use project files will stay
supported.