tools/r2plugin: change args to satisfy Windows path convention
There is a constraint in naming files on Windows. Each file
cannot contain character from `[?/\*":<>|]`. This commit
provides change in r2plugin args that did result in
such special characters to be used as directory names and thus
in failing tests. To not use such characters different
ways were used to achive same behavior that was tested.

This is a workaround - real solulution would require OS specific
sanitization of paths that is not required right now with limited
use cases.