integration/structure-local: do not recompile on macOS
macOS 10.15 does not contain support for 32 bit applications.
This means that for recompilation is used 64 bit compiler
that produces 64 bit binary. Because word size of original
binary was 32 bits result C file contains constructions
that can be run only on 32 bit architectures.

With merge of improvements in structure analysis that is being
worked on we can enable recompilation and run analysis
of these tests.