integration/nested-for-loop: temporary fix of a bug
This bug seems to be rooted in clang python bindings or the way
framework is using it. The increment in the loop is present,
however, it is not parsed and provided to the test.

This should be fixed by fixing clang/framework. Right now
it is fixed by providing loop that is being parsed on macOS.

This bug is not relevant for other system than macOS.