Modify test to agree with RetDec using patched Capstone 4.0.2
Capstone version 4.0.2 has a bug when disassembling a powerpc instruction
with a signed 16-bit immediate.
See https://github.com/capstone-engine/capstone/issues/1746 and
https://github.com/capstone-engine/capstone/issues/1746#issuecomment-1186559582.

RetDec PR https://github.com/avast/retdec/pull/1086 moved RetDec up to
version 4.0.2 of Capstone, but adds a patch to fix this bug. The
fix includes not printing the immediates in hex (a later version
of Capstone that fixes this problem no long prints the immediates in hex).
This change fixes a test to agree with the new behavior.