Add and modify tests for prototypes of dynamically-linked functions
Add tests that fail if dynamically-linked functions are not given
prototypes. These call "operator new", that is _Znwj or _Xnwm.
Without the prototype they will be assumed to return "int" which
cannot be dereferenced.  A companion change modifies RetDec to emit
prototypes for dynamically-linked functions.

This change also includes changes to other tests that were
affected by this change.